Amelioration pour finalisation 3

+++ frontend-app/patients.go
@@ -9,14 +9,12 @@ import (
 
 // Affichage de la liste
 func (a *App) GetPatients() []Patient {
-	rows, err := db.Query("SELECT id, nom, prenom, sexe, telephone, date_naissance, adresse, motif FROM patients")
+rows, err := db.Query("SELECT id, nom, prenom, sexe, telephone, date_naissance, adresse, motif FROM patients ORDER BY id DESC")
 	if err != nil {
 		fmt.Println("Erreur recuperation patients:", err)
 		return []Patient{}
 	}
-
 	defer rows.Close() //ferme la connexion aux resultats a la fin de la fonction
-
 	//recuperation des données de chaque ligne de patient
 	var patients []Patient = []Patient{} // Initialisé vide pour éviter `null` en JS
 	for rows.Next() {
@@ -34,6 +32,10 @@ func (a *App) AddPatient(nom, prenom, sexe, telephone, dateNaissance, adresse, m
 
 	if !matched {
 		return "Le numéro de téléphone doit contenir exactement 9 chiffres."
+	}
+
+	if sexe != "M" && sexe != "F" {
+		return "Le sexe doit être 'M' ou 'F'."
 	}
 
 	var count int
@@ -54,10 +56,13 @@ func (a *App) AddPatient(nom, prenom, sexe, telephone, dateNaissance, adresse, m
 
 // Mises à jour d'un patient pour modification
 func (a *App) UpdatePatient(id int, nom, prenom, sexe, telephone, dateNaissance, adresse, motif string) string {
+	
 	matched, _ := regexp.MatchString(`^\d{8,15}$`, telephone)
-
 	if !matched {
 		return "Le numéro de téléphone doit contenir exactement 9 chiffres."
+	}
+	if sexe != "M" && sexe != "F" {
+		return "Le sexe doit être 'M' ou 'F'."
 	}
 
 	var count int
@@ -65,7 +70,6 @@ func (a *App) UpdatePatient(id int, nom, prenom, sexe, telephone, dateNaissance,
 	if count > 0 {
 		return "Ce numéro de téléphone est déjà utilisé par un autre patient."
 	}
-
 	_, err := db.Exec(
 		"UPDATE patients SET nom=?, prenom=?, sexe=?, telephone=?, date_naissance=?, adresse=?, motif=? WHERE id=?",
 		nom, prenom, sexe, telephone, dateNaissance, adresse, motif, id,
@@ -78,6 +82,16 @@ func (a *App) UpdatePatient(id int, nom, prenom, sexe, telephone, dateNaissance,
 
 // Suppression definitive d'un patient grace a son ID
 func (a *App) DeletePatient(id int) string {
+	var countRdv int
+	db.QueryRow("SELECT COUNT(*) FROM rendez_vous WHERE patient_id = ?", id).Scan(&countRdv)
+	if countRdv > 0 {
+		return "Impossible de supprimer ce patient : il a des rendez-vous enregistrés dans l'historique."
+	}
+	var countConsult int
+	db.QueryRow("SELECT COUNT(*) FROM consultations WHERE patient_id = ?", id).Scan(&countConsult)
+	if countConsult > 0 {
+		return "Impossible de supprimer ce patient : il a des consultations enregistrées dans l'historique."
+	}
 	_, err := db.Exec("DELETE FROM patients WHERE id=?", id)
 	if err != nil {
 		return "Erreur: " + err.Error()
